+++ array/zfparray4.h
@@ -199,9 +199,9 @@ public:
     const ptrdiff_t sz = static_cast<ptrdiff_t>(nx * ny);
     const ptrdiff_t sw = static_cast<ptrdiff_t>(nx * ny * nz);
     size_t block_index = 0;
-    for (size_t l = 0; l < bw; l++, p += 4 * sz * (nz - bz))
-      for (size_t k = 0; k < bz; k++, p += 4 * sy * (ny - by))
-        for (size_t j = 0; j < by; j++, p += 4 * sx * (nx - bx))
+    for (size_t l = 0; l < bw; l++, p += 4 * sz * ptrdiff_t(nz - bz))
+      for (size_t k = 0; k < bz; k++, p += 4 * sy * ptrdiff_t(ny - by))
+        for (size_t j = 0; j < by; j++, p += 4 * sx * ptrdiff_t(nx - bx))
           for (size_t i = 0; i < bx; i++, p += 4)
             cache.get_block(block_index++, p, sx, sy, sz, sw);
   }
@@ -218,9 +218,9 @@ public:
     const ptrdiff_t sz = static_cast<ptrdiff_t>(nx * ny);
     const ptrdiff_t sw = static_cast<ptrdiff_t>(nx * ny * nz);
     size_t block_index = 0;
-    for (size_t l = 0; l < bw; l++, p += 4 * sz * (nz - bz))
-      for (size_t k = 0; k < bz; k++, p += 4 * sy * (ny - by))
-        for (size_t j = 0; j < by; j++, p += 4 * sx * (nx - bx))
+    for (size_t l = 0; l < bw; l++, p += 4 * sz * ptrdiff_t(nz - bz))
+      for (size_t k = 0; k < bz; k++, p += 4 * sy * ptrdiff_t(ny - by))
+        for (size_t j = 0; j < by; j++, p += 4 * sx * ptrdiff_t(nx - bx))
           for (size_t i = 0; i < bx; i++, p += 4)
             cache.put_block(block_index++, p, sx, sy, sz, sw);
   }
